:art: ::  MainActivity 이동 시키기
