+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Hiv2"
         tools:targetApi="31">
-        <activity
-            android:name=".MainActivity"
-            android:exported="true"
-            android:label="Hi-v2">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
     </application>
 
 </manifest>
