+++ presentation/src/main/java/team/msg/presentation/view/MainActivity.kt
@@ -1,0 +1,16 @@
+package team.msg.presentation.view
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            LoginScreen()
+        }
+    }
+}
